Add default Exception to events
Add default exception for handling events such that we can at least display the error message.
This is dirty fix for solving this issue, as all exepction are caught like this.
We should investigate which exceptions could actually occur, and only use them.
Others should be caught by the users themselves.